src/priority_queue/queue.go
package priority_queue

type Item interface {
	lt(other interface{}) bool
}

type PriorityQueue struct {
	items []*Item
}

func (pq *PriorityQueue) Len() int {
	return len(pq.items)
}

func (pq *PriorityQueue) Less(i, j int) bool {
	first := *pq.items[i]
	second := *pq.items[j]
	return first.lt(second)
}

func (pq *PriorityQueue) Swap(i, j int) {
	pq.items[i], pq.items[j] = pq.items[j], pq.items[i]
}

func (pq *PriorityQueue) Push(x interface{}) {
	pq.items = append(pq.items, x.(*Item))
}

func (pq *PriorityQueue) Pop() interface{} {
	old := pq.items
	n := len(old)
	item := old[n-1]
	pq.items = old[0 : n-1]
	return item
}

func (pq *PriorityQueue) Peek() interface{} {
	return pq.items[0]
}
